NEW COMMIT WITH PROPER STRUCTURE
