+++ src/main/java/com/globalitgeeks/examninja/config/SwaggerConfig.java
@@ -1,0 +1,23 @@
+package com.globalitgeeks.examninja.config;
+import io.swagger.v3.oas.models.ExternalDocumentation;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info().title("Exam Ninja API")
+                        .description("API documentation for the Exam Ninja application")
+                        .version("v1.0")
+                        .license(new License().name("Apache 2.0").url("http://springdoc.org")))
+                .externalDocs(new ExternalDocumentation()
+                        .description("Exam Ninja Documentation")
+                        .url("https://exam-ninja.docs"));
+    }
+}
