+++ src/main/java/com/globalitgeeks/examninja/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.globalitgeeks.examninja.usermanagement.config;
+package com.globalitgeeks.examninja.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
