+++ src/main/java/com/globalitgeeks/examninja/dto/ApiResponse.java
@@ -1,0 +1,11 @@
+package com.globalitgeeks.examninja.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ApiResponse{
+    private String status;
+    private String message;
+}
