+++ src/main/java/com/globalitgeeks/examninja/dto/CustomisedErrorResponse.java
@@ -1,0 +1,12 @@
+package com.globalitgeeks.examninja.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class CustomisedErrorResponse {
+    private String message;
+}
