+++ src/main/java/com/globalitgeeks/examninja/dto/ErrorResponse.java
@@ -1,0 +1,15 @@
+package com.globalitgeeks.examninja.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Map;
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorResponse {
+    private String message;
+    private Map<String, String> error;
+
+}
