+++ src/main/java/com/globalitgeeks/examninja/exception/InvalidPasswordException.java
@@ -1,0 +1,7 @@
+package com.globalitgeeks.examninja.exception;
+
+public class InvalidPasswordException extends RuntimeException {
+    public InvalidPasswordException(String message) {
+        super(message);
+    }
+}
