+++ src/main/java/com/globalitgeeks/examninja/exception/UserAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.globalitgeeks.examninja.exception;
+
+public class UserAlreadyExistsException extends RuntimeException {
+    public UserAlreadyExistsException(String message) {
+        super(message);
+    }
+}
