+++ src/main/java/com/globalitgeeks/examninja/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.globalitgeeks.examninja.exception;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}
