+++ src/main/java/com/globalitgeeks/examninja/model/User.java
@@ -1,0 +1,23 @@
+package com.globalitgeeks.examninja.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Entity
+@Table(name = "users")
+@Data
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String firstName;
+    private String lastName;
+
+    @Column(unique = true) // Ensure email is unique
+    private String email;
+
+    private String password;
+}
+
