+++ src/main/java/com/globalitgeeks/examninja/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.globalitgeeks.examninja.repository;
+
+import com.globalitgeeks.examninja.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
