--- src/main/java/com/globalitgeeks/examninja/usermanagement/dto/ApiResponse.java
@@ -1,11 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ApiResponse{
-    private String status;
-    private String message;
-}
