--- src/main/java/com/globalitgeeks/examninja/usermanagement/dto/CustomisedErrorResponse.java
@@ -1,12 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CustomisedErrorResponse {
-    private String message;
-}
