--- src/main/java/com/globalitgeeks/examninja/usermanagement/dto/ErrorResponse.java
@@ -1,15 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorResponse {
-    private String message;
-    private Map<String, String> error;
-
-}
