--- src/main/java/com/globalitgeeks/examninja/usermanagement/exception/InvalidPasswordException.java
@@ -1,7 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.exception;
-
-public class InvalidPasswordException extends RuntimeException {
-    public InvalidPasswordException(String message) {
-        super(message);
-    }
-}
