--- src/main/java/com/globalitgeeks/examninja/usermanagement/exception/UserAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.exception;
-
-public class UserAlreadyExistsException extends RuntimeException {
-    public UserAlreadyExistsException(String message) {
-        super(message);
-    }
-}
