--- src/main/java/com/globalitgeeks/examninja/usermanagement/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.exception;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
