--- src/main/java/com/globalitgeeks/examninja/usermanagement/model/User.java
@@ -1,30 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.persistence.Column;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "users")
-@Data
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String firstName;
-    private String lastName;
-
-    @Column(unique = true) // Ensure email is unique
-    private String email;
-
-    private String password;
-}
-
