--- src/main/java/com/globalitgeeks/examninja/usermanagement/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.globalitgeeks.examninja.usermanagement.repository;
-
-import com.globalitgeeks.examninja.usermanagement.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
